add the ability to delete tweets older than a certain date

+++ csv.go
@@ -5,11 +5,17 @@ import (
 	"fmt"
 	"os"
 	"strconv"
+	"time"
 
 	"github.com/ChimeraCoder/anaconda"
 )
 
 func csv(api *anaconda.TwitterApi) error {
+	t, err := date(*olderThan)
+	if err != nil {
+		return err
+	}
+
 	f, err := os.Open(*csvFile)
 	if err != nil {
 		return err
@@ -29,6 +35,16 @@ func csv(api *anaconda.TwitterApi) error {
 		id, err := strconv.ParseInt(rec[0], 10, 64)
 		if err != nil {
 			return err
+		}
+
+		ti, err := time.Parse("2006-01-02 15:04:05 -0700", rec[3])
+		if err != nil {
+			return err
+		}
+
+		// skip if newer than t
+		if ti.After(t) {
+			continue
 		}
 
 		_, err = api.DeleteTweet(id, true)
